Item Catalog

+++ catalog/application.py
@@ -8,7 +8,7 @@ import httplib2
 from oauth2client.client import FlowExchangeError
 from oauth2client.client import flow_from_clientsecrets
 from flask import session as login_session
-from project_database import Base, Mineral, Item
+from project_database import Base, Mineral, Item, User
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy import create_engine
 """
@@ -17,12 +17,14 @@ Created on Mon Nov 26 16:33:42 2018
 @author: Nicolas
 """
 
-from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
+from flask import Flask, render_template, request, redirect,
+url_for, flash, jsonify
 
 app = Flask(__name__)
 
 
-# Declare my client_id
+""" Declare my client_id"""
+
 CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())[
     'web']['client_id']
 engine = create_engine('sqlite:///mineralsitems.db?check_same_thread=False')
@@ -32,8 +34,8 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-# Create a state token to prevent request forgery
-# Store it in the session for later validation
+"""Create a state token to prevent request forgery
+   Store it in the session for later validation"""
 
 
 @app.route('/gconnect', methods=['POST'])
@@ -56,7 +58,8 @@ def gconnect():
         return response
     # Check that the access token is valid.
     access_token = credentials.access_token
-    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
     h = httplib2.Http()
     result = json.loads(h.request(url, 'GET')[1])
     # If there was an error in the access token info, abort.
@@ -90,16 +93,20 @@ def gconnect():
     print(credentials.access_token)
     login_session['gplus_id'] = gplus_id
 
-    #Get user info
+    # Get user info
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
-    #print(answer)
     data = answer.json()
-    print(data.keys()) #
     login_session['username'] = data["name"]
     login_session['picture'] = data["picture"]
     login_session['email'] = data["email"]
+
+    # see if user exists, if it don't make a new one
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
 
     output = ''
     output += '<h1>Welcome,'
@@ -108,14 +115,30 @@ def gconnect():
 
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height300px;border-radius: 150px;' \
+        '-webkit-border-radius:150px;-moz-border-radius: 150px;"> '
 
     flash("you are now logged in as %s" % login_session["username"])
     print("done!")
     return output
 
+"""logout user"""
 
-#DISCONNECT - Revoke a current user's token and reset their login_session.
+
+@app.route('/logout', methods=['POST'])
+def logout():
+    # Disconnect based on provider
+
+    if login_session.get('provider') == 'google':
+        return gdisconnect()
+    else:
+        response = make_response(json.dumps({'state': 'notConnected'}),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+""" DISCONNECT - Revoke a current user's token and reset their login_session.
+   """
 
 
 @app.route("/gdisconnect")
@@ -134,7 +157,7 @@ def gdisconnect():
     result = h.request(url, 'GET')[0]
 
     if result['status'] == '200':
-        #Reset the user's session.
+        # Reset the user's session.
         del login_session['credentials']
         del login_session['gplus_id']
         del login_session['username']
@@ -145,19 +168,23 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     else:
-        #For whatever reason, the given token was Invalid
-        response = make_response(json.dumps('Failed to revoke token for given user.'), 400)
+        # For whatever reason, the given token was Invalid
+        response = make_response(json.dumps('Failed to revoke '
+                                 'token for given user.'), 400)
         response.headers['Content-Type'] = 'application/json'
         return response
 
 
 @app.route('/login')
 def showLogin():
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
     login_session['state'] = state
     return render_template('login.html', STATE=state)
 
-# Shows the different types of minerals 
+""" Shows the different types of minerals """
+
+
 @app.route('/')
 @app.route('/minerals/')
 def showMinerals():
@@ -165,13 +192,16 @@ def showMinerals():
     return render_template('minerals.html', minerals=minerals)
 
 
-#Create a new mineral. For this step is necessary to authenticate first 
+""" Create a new mineral. For this step is necessary to authenticate first """"
+
+
 @app.route('/minerals/new/', methods=['GET', 'POST'])
 def newMineral():
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
-        new_mineral = Mineral(name=request.form['name'])
+        new_mineral = Mineral(name=request.form['name'],
+                              user_id=login_session['user_id'])
         session.add(new_mineral)
         session.commit()
         flash('New Mineral Created Successfully')
@@ -179,67 +209,104 @@ def newMineral():
     else:
         return render_template('NewMineral.html')
 
-# Edit a Mineral name. This route can be used in order to correct misspelled names of minerals or update names.
-@app.route('/minerals/<int:mineral_id>/edit/', methods=['GET','POST'])
+""" Edit a Mineral name. This route can be used in order to correct misspelled
+ names of minerals or update names."""
+
+
+@app.route('/minerals/<int:mineral_id>/edit/', methods=['GET', 'POST'])
 def editMineral(mineral_id):
-    edit_mineral = session.query(Mineral).filter_by(id= mineral_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    edit_mineral = session.query(Mineral).filter_by(id=mineral_id).one()
     if request.method == 'POST':
         if request.form['name']:
             edit_mineral.name = request.form['name']
         session.add(edit_mineral)
         session.commit()
         flash('Mineral edited successfully')
-        return redirect(url_for('showItems',mineral_id=mineral_id))
+        return redirect(url_for('showItems', mineral_id=mineral_id))
     else:
-        return render_template('editmineral.html', mineral_id=mineral_id, i= edit_mineral)
+        return render_template('editmineral.html', mineral_id=mineral_id,
+                               i=edit_mineral)
 
-# Delete a Mineral. This route can be used to delete a Mineral that has incorrect name, or a mineral that was added before. 
-@app.route('/minerals/<int:mineral_id>/delete/', methods=['GET','POST'])
+""" Delete a Mineral. This route can be used to delete a
+ Mineral that has incorrect name, or a mineral that was added before."""
+
+
+@app.route('/minerals/<int:mineral_id>/delete/', methods=['GET', 'POST'])
 def deleteMineral(mineral_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     delete_mineral = session.query(Mineral).filter_by(id=mineral_id).one()
-    if request.method== 'POST':
+    if request.method == 'POST':
         session.delete(delete_mineral)
         session.commit()
         flash('Mineral successfully deleted')
         return redirect(url_for('showMinerals'))
     else:
-        return render_template('DeleteMineral.html', mineral_id = mineral_id, i = delete_mineral)
+        return render_template('DeleteMineral.html',
+                               mineral_id=mineral_id, i=delete_mineral)
 
-# Show information of the differents items that belong to a group of mineral 
+""" Show information of the differents items that belong to a group
+    of mineral"""
+
+
 @app.route('/minerals/<int:mineral_id>/items/')
 def showItems(mineral_id):
     mineral = session.query(Mineral).filter_by(id=mineral_id).one()
     items = session.query(Item).filter_by(mineral_id=mineral_id)
-    return render_template('items.html', mineral_id=mineral_id, mineral=mineral, items=items)
+    return render_template('items.html', mineral_id=mineral_id,
+                           mineral=mineral, items=items)
 
-# Show item information such as price, origin, colour, description
+""" Show item information such as price, origin, colour, description"""
+
+
 @app.route('/minerals/<int:mineral_id>/items/<int:item_id>/')
 def showItemInformation(mineral_id, item_id):
     mineral = session.query(Mineral).filter_by(id=mineral_id).one()
     item = session.query(Item).filter_by(id=item_id)
-    return render_template('itemInfo.html',mineral=mineral, mineral_id=mineral_id, item_id=item_id, item=item)
+    return render_template('itemInfo.html', mineral=mineral,
+                           mineral_id=mineral_id, item_id=item_id, item=item)
 
-# Create a new item within a category. Firt authenticate in order to proceed with the new item generation
+""" Create a new item within a category. Firt authenticate in order to
+ proceed with the new item generation"""
+
+
 @app.route('/minerals/<int:mineral_id>/items/new/', methods=['GET', 'POST'])
 def newItem(mineral_id):
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
-        new_item = Item(name=request.form['name'], origin=request.form['origin'], colour=request.form['colour'], price=request.form['price'],
-                        hardness=request.form['hardness'], description=request.form['description'], mineral_id=mineral_id)
+        new_item = Item(name=request.form['name'],
+                        origin=request.form['origin'],
+                        colour=request.form['colour'],
+                        price=request.form['price'],
+                        hardness=request.form['hardness'],
+                        description=request.form['description'],
+                        mineral_id=mineral_id,
+                        user_id=mineral.user_id)
         session.add(new_item)
         session.commit()
         return redirect(url_for('showItems', mineral_id=mineral_id))
     else:
         return render_template('newItem.html', mineral_id=mineral_id)
 
-# Edit information of the item. The user must be authenticate in the web app in order to proceed with the change of item properties.
-@app.route('/minerals/<int:mineral_id>/items/<int:item_id>/edit/', methods=['GET', 'POST'])
+""" Edit information of the item. The user must be authenticate in the
+ web app in order to proceed with the change of item properties."""
+
+
+@app.route('/minerals/<int:mineral_id>/items/<int:item_id>/edit/',
+           methods=['GET', 'POST'])
 def editItem(mineral_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
     mineral = session.query(Mineral).filter_by(id=mineral_id).one()
     edit_item = session.query(Item).filter_by(id=item_id).one()
+    if edit_item.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You\
+         are not authorized to edit this item.\
+          Please create your own item in order\
+           to edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             edit_item.name = request.form['name']
@@ -272,38 +339,77 @@ def editItem(mineral_id, item_id):
         session.commit()
 
         flash('Item successfully edited')
-        return redirect(url_for('showItems', mineral_id=mineral_id, item_id=item_id))
+        return redirect(url_for('showItems', mineral_id=mineral_id,
+                                item_id=item_id))
     else:
-        return render_template('editItem.html', mineral_id=mineral_id, item_id=item_id, item=edit_item)
+        return render_template('editItem.html', mineral_id=mineral_id,
+                               item_id=item_id, item=edit_item)
 
-# Delete a specific item. The user must be authenticate in the web app in order to remove items within a category. 
-@app.route('/minerals/<int:mineral_id>/items/<int:item_id>/delete/', methods=['GET', 'POST'])
+""" Delete a specific item. The user must be authenticate in the web app
+ in order to remove items within a category."""
+
+
+@app.route('/minerals/<int:mineral_id>/items/<int:item_id>/delete/',
+           methods=['GET', 'POST'])
 def deleteItem(mineral_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
     item_to_delete = session.query(Item).filter_by(id=item_id).one()
+    if item_to_delete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You\
+         are not authorized to edit this item.\
+          Please create your own item in order\
+           to edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(item_to_delete)
         session.commit()
         flash('Item Successfully deleted')
         return redirect(url_for('showItems', mineral_id=mineral_id))
     else:
-        return render_template('deleteItem.html', mineral_id=mineral_id,  i=item_to_delete)
+        return render_template('deleteItem.html',
+                               mineral_id=mineral_id,  i=item_to_delete)
 
 
-# Return serialized information of the list of minerals in the web app
+""" Return serialized information of the list of minerals in the web app"""
+
+
 @app.route('/minerals/JSON')
 def minerals_to_json():
     minerals = session.query(Mineral).all()
     return jsonify(minerals=[i.serialize for i in minerals])
 
-# Return serialized information of properties of the items within a specific group of minerals defined in /minerals route.
+""" Return serialized information of properties of the items within a
+ specific group of minerals defined in /minerals route."""
+
+
 @app.route('/minerals/<int:mineral_id>/items/<int:item_id>/JSON')
 def menuitem(mineral_id, item_id):
     minerals = session.query(Mineral).filter_by(id=mineral_id).one()
     item = session.query(Item).filter_by(mineral_id=mineral_id, id=item_id)
     return jsonify(item=[i.serialize for i in item])
 
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email)
+        return user.id
+    except:
+        return None
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'],
+                   email=login_session['email'],
+                   picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
 
 if __name__ == '__main__':
     app.secret_key = 'supersecretkey'
